♻️ refactor: react-query core 파일 이동 - /application/hooks/api/core -> /api/core - useCoreInfiniteQuery 추후 리팩터링 시 삭제 예정
