+++ src/api/core/useCoreInfiniteQuery.ts
