+++ src/api/core/useSuspensedQuery.ts
@@ -1,13 +1,41 @@
-import type { QueryFunction, QueryKey } from "@tanstack/react-query";
+import type {
+  QueryFunction,
+  QueryKey,
+  UseQueryOptions,
+  UseQueryResult,
+} from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
 
-import type {
-  BaseSuspendedUseQueryResult,
-  SuspendedUseQueryOptions,
-  SuspendedUseQueryResultOnIdle,
-  SuspendedUseQueryResultOnSuccess,
-} from "@/application/hooks/api/core/types";
+export interface BaseSuspendedUseQueryResult<TData>
+  extends Omit<
+    UseQueryResult<TData, never>,
+    "data" | "status" | "error" | "isLoading" | "isError" | "isFetching"
+  > {
+  data: TData;
+  status: "success" | "idle";
+}
 
+export type SuspendedUseQueryResultOnSuccess<TData> = BaseSuspendedUseQueryResult<TData> & {
+  status: "success";
+  isSuccess: true;
+  isIdle: false;
+};
+export type SuspendedUseQueryResultOnIdle<TData> = BaseSuspendedUseQueryResult<TData> & {
+  status: "idle";
+  isSuccess: false;
+  isIdle: true;
+};
+
+export type SuspendedUseQueryOptions<
+  TQueryFnData = unknown,
+  TError = unknown,
+  TData = TQueryFnData,
+  TQueryKey extends QueryKey = QueryKey,
+> = Omit<UseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, "suspense">;
+
+export type QuerySelectOption<Result, API extends (...args: any) => any> = (
+  data: Awaited<ReturnType<API>>,
+) => Result;
 /**
  * @desc suspense 사용 시 non-nullable data 이도록 wrapping
  * @link https://github.com/toss/slash/blob/main/packages/react/react-query/src/hooks/useSuspendedQuery.ts
