+++ src/features/common/components/InfiniteMemeList.tsx
@@ -2,8 +2,8 @@ import type { UseInfiniteQueryResult } from "@tanstack/react-query";
 import { useQueryClient } from "@tanstack/react-query";
 import { useCallback, useMemo } from "react";
 
+import { CORE_QUERY_KEY } from "@/api/core";
 import { prefetchCollectionCheck } from "@/application/hooks";
-import { CORE_QUERY_KEY } from "@/application/hooks/api/core/queryKey";
 import { QUERY_KEYS } from "@/application/hooks/api/meme/queryKey";
 import { Masonry } from "@/common/components/Masonry";
 import { renderMemeItemSkeletons } from "@/common/components/Skeleton";
